checkTimout and randomAssigner impl

+++ src/main/java/com/marafone/marafone/game/event/incoming/CardSelectEvent.java
@@ -5,5 +5,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CardSelectEvent {
-    public int cardId;
+    public long cardId;
 }
